Реорганизация store. Именен reducer data=>films
