--- src/store/selectors.js
@@ -1,27 +1,0 @@
-
-// export const selectFilms = (state) => state[NameSpace.DATA].films;
-// export const selectFavoriteFilms = (state) => state[NameSpace.DATA].favoriteFilms;
-// export const selectFilm = (state) => state[NameSpace.DATA].film;
-// export const selectPromoFilm = (state) => state[NameSpace.DATA].promoFilm;
-// export const selectReviews = (state) => state[NameSpace.DATA].reviews;
-// export const selectIsFilmLoaded = (id) => (state) => selectFilm(state).id === id;
-
-// export const selectCurrentGenre = (state) => state[NameSpace.CATALOG].currentGenre;
-// export const selectShownFilmsCount = (state) => state[NameSpace.CATALOG].shownFilmsCount;
-
-// export const selectAuthorizationStatus = (state) => state[NameSpace.USER].authorizationStatus;
-// export const selectUserInfo = (state) => state[NameSpace.USER].userInfo;
-// export const selectUserRequestErrorCode = (state) => state[NameSpace.USER].errorCode;
-// export const selectUserRequestStatus = (state) => state[NameSpace.USER].requestStatus;
-// export const selectIsUserRequested = (state) => selectUserRequestStatus(state) === RequestStatus.REQUESTED;
-// export const selectIsUserRequestFailed = (state) => selectUserRequestStatus(state) === RequestStatus.REQUEST_FAILED;
-
-// export const selectReviewRequestStatus = (state) => state[NameSpace.REVIEW].requestStatus;
-// export const selectIsReviewPostRequested = (state) => selectReviewRequestStatus(state) === RequestStatus.REQUESTED;
-// export const selectIsReviewPostFailed = (state) => selectReviewRequestStatus(state) === RequestStatus.REQUEST_FAILED;
-
-// export const selectIsUserLogged = (state) => {
-//   return selectAuthorizationStatus(state) === AuthorizationStatus.AUTH;
-// };
-
-
