+++ src/store/user/user.js
@@ -20,7 +20,7 @@ const initialState = {
 
 // Reducer
 
-export const user = (state = initialState, action) => {
+export const userReducer = (state = initialState, action) => {
   switch (action.type) {
     case ActionType.SET_AUTHORIZATION_STATUS:
       return extend(state, {authorizationStatus: action.payload});
@@ -49,9 +49,9 @@ export const setAuthorizationStatus = (status) => ({
   payload: status,
 });
 
-export const loadUser = (userInfo) => ({
+export const loadUser = (user) => ({
   type: ActionType.LOAD_USER,
-  payload: userInfo,
+  payload: user,
 });
 
 export const userRequested = () => ({
